requirements-api.txt
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
mcp>=1.26.0
PyJWT[crypto]>=2.10.1
pypdf>=5.4.0
python-docx>=1.1.2
python-pptx>=1.0.2
